:truck: Rename KeyContext.tsx to key.context.tsx
