+++ screens/contacts.screen.tsx
@@ -5,7 +5,7 @@ import { Layout } from '../components/layout.component';
 import { NoContent } from '../components/no-content.component';
 import { RefreshControl } from '../components/refresh-control.component';
 import { Spinner } from '../components/spinner.component';
-import { KeyContext } from '../context/KeyContext';
+import { KeyContext } from '../context/key.context';
 import { Profile, ReallifeRPGService } from '../services/realliferpg.service';
 
 export const Contacts: React.FC = () => {
